refactor: isolate supabase server and client helpers
